Make authenticated dashboard the home route
